feat: add createNode() and createLink() factories
to match DAGNode and DAGLink constructors of old
