+++ README.md
@@ -5,7 +5,7 @@ An implementation of the [DAG-PB spec](https://github.com/ipld/specs/blob/master
 ## Example
 
 ```js
-import CID from 'multiformats/cid'
+import { CID } from 'multiformats/cid'
 import { sha256 } from 'multiformats/hashes/sha2'
 import * as dagPB from '@ipld/dag-pb'
 
@@ -47,7 +47,7 @@ The DAG-PB encoding is very strict about the Data Model forms that are passed in
 Due to this strictness, a `prepare()` function is made available which simplifies construction and allows for more flexible input forms. Prior to encoding objects, call `prepare()` to receive a new object that strictly conforms to the schema.
 
 ```js
-import CID from 'multiformats/cid'
+import { CID } from 'multiformats/cid'
 import { prepare } from '@ipld/dag-pb'
 
 console.log(prepare({ Data: 'some data' }))
@@ -65,6 +65,61 @@ Some features of `prepare()`:
 * Ensuring that properties are of the correct type (link `Name` is a `string` and `Tsize` is a `number`)
 * `Links` array is always present, even if empty
 * `Links` array is properly sorted
+
+## `createNode()` & `createLink()`
+
+These utility exports are available to make transition from the older [ipld-dag-pb](https://github.com/ipld/js-ipld-dag-pb) library which used `DAGNode` and `DAGLink` objects with constructors. `createNode()` mirrors the `new DAGNode()` API while `createLink()` mirrors `new DAGLink()` API.
+
+* `createNode(data: Uint8Array, links: PBLink[]|void): PBNode`: create a correctly formed `PBNode` object from a `Uint8Array` and an optional array of correctly formed `PBLink` objects. The returned object will be suitable for passing to `encode()` and using `prepare()` on it should result in a noop.
+* `createLink(name: string, size: number, cid: CID): PBLink`: create a correctly formed `PBLink` object from a name, size and CID. The returned object will be suitable for attaching to a `PBNode`'s `Links` array, or in an array for the second argument to `createNode()`.
+
+```js
+import { CID, bytes } from 'multiformats'
+import * as Block from 'multiformats/block'
+import { sha256 as hasher } from 'multiformats/hashes/sha2'
+import * as codec from '@ipld/dag-pb'
+
+const { createLink, createNode } = codec
+
+async function run () {
+  const cid1 = CID.parse('QmWDtUQj38YLW8v3q4A6LwPn4vYKEbuKWpgSm6bjKW6Xfe')
+  const cid2 = CID.parse('bafyreifepiu23okq5zuyvyhsoiazv2icw2van3s7ko6d3ixl5jx2yj2yhu')
+
+  const links = [createLink('link1', 100, cid1), createLink('link2', 200, cid2)]
+  const value = createNode(Uint8Array.from([0, 1, 2, 3, 4]), links)
+  console.log(value)
+
+  const block = await Block.encode({ value, codec, hasher })
+  console.log(block.cid)
+  console.log(`Encoded: ${bytes.toHex(block.bytes).replace(/(.{80})/g, '$1\n         ')}`)
+}
+
+run().catch((err) => console.error(err))
+```
+
+Results in:
+
+```
+{
+  Data: Uint8Array(5) [ 0, 1, 2, 3, 4 ],
+  Links: [
+    {
+      Hash: CID(QmWDtUQj38YLW8v3q4A6LwPn4vYKEbuKWpgSm6bjKW6Xfe),
+      Name: 'link1',
+      Tsize: 100
+    },
+    {
+      Hash: CID(bafyreifepiu23okq5zuyvyhsoiazv2icw2van3s7ko6d3ixl5jx2yj2yhu),
+      Name: 'link2',
+      Tsize: 200
+    }
+  ]
+}
+CID(bafybeihsp53wkzsaif76mjv564cawzqyjwianosamlvf6sht2m25ttyxiy)
+Encoded: 122d0a2212207521fe19c374a97759226dc5c0c8e674e73950e81b211f7dd3b6b30883a08a511205
+         6c696e6b31186412300a2401711220a47a29adb950ee698ae0f272019ae902b6aa06ee5f53bc3da2
+         ebea6fac27583d12056c696e6b3218c8010a050001020304
+```
 
 ## License
